Gut the project like a fish
In the midst of developing the original concept for hero-hero, I started
working on a different project (a UI for the RPG battle system I
developed a few months ago) using this project as a dev ground. That's a
mistake; so, I'm moving all the offending code to a different
repository.