Handle slightly more complex item usage
The sword item can be picked up just by clicking on it, and it will not
be released by releasing the mouse button of the arm. Hold down 'E' and
press the mouse button corresponding to the arm holding the sword and it
will be dropped.

TODO:
Sometimes, when the sword is dropped, it falls through the floor. Not
sure what causes that, but will need to be fixed.